refactory remoting socket handling.

+++ src/ECommon/Remoting/Exceptions/RemotingTimeoutException.cs
@@ -4,7 +4,7 @@ namespace ECommon.Remoting.Exceptions
 {
     public class RemotingTimeoutException : Exception
     {
-        public RemotingTimeoutException(string address, RemotingRequest request, int timeoutMillis)
+        public RemotingTimeoutException(string address, RemotingRequest request, long timeoutMillis)
             : base(string.Format("Wait response on the channel <{0}> timeout, request:{1}, timeoutMillis:{2}ms", address, request, timeoutMillis))
         {
         }
